02: Loosely couple the Components
