+++ src/CMakeLists.txt
@@ -40,6 +40,8 @@ endif()
 
 include(cmake/shared-sources.cmake)
 include(cmake/sources.cmake)
+include(webextension/CMakeLists.txt)
+
 
 target_link_libraries(mozillavpn PRIVATE
     shared-sources
@@ -67,5 +69,9 @@ endif()
 
 add_subdirectory(ui)
 target_link_libraries(mozillavpn PRIVATE mozillavpn-ui mozillavpn-uiplugin)
+qt6_add_qml_module(mozillavpn
+  URI Mozilla.Shared.qmlcomponents
+  VERSION 1.0
+)
 
 qt_finalize_target(mozillavpn)
