+++ src/cmake/sources.cmake
@@ -212,10 +212,6 @@ if(NOT CMAKE_CROSSCOMPILING)
      target_sources(mozillavpn-sources INTERFACE
         ${CMAKE_CURRENT_SOURCE_DIR}/systemtraynotificationhandler.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/systemtraynotificationhandler.h
-        ${CMAKE_CURRENT_SOURCE_DIR}/webextension/connection.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/webextension/connection.h
-        ${CMAKE_CURRENT_SOURCE_DIR}/webextension/server.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/webextension/server.h
        )
 endif()
