+++ src/commands/commandui.cpp
@@ -92,6 +92,7 @@
 
 #ifdef MVPN_WEBEXTENSION
 #  include "webextension/server.h"
+#  include "webextension/vpnadapter.h"
 #endif
 
 #include <QApplication>
@@ -394,7 +395,7 @@ int CommandUI::run(QStringList& tokens) {
 #endif
 
 #ifdef MVPN_WEBEXTENSION
-    WebExtension::Server extensionServer;
+    WebExtension::Server extensionServer(new WebExtension::VPNAdapter(qApp));
     QObject::connect(vpn.controller(), &Controller::readyToQuit,
                      &extensionServer, &WebExtension::Server::close);
 #endif
