+++ src/webextension/connection.cpp
@@ -5,13 +5,14 @@
 #include "webextension/connection.h"
 
 #include <QHostAddress>
+#include <QIODevice>
 #include <QJsonArray>
 #include <QJsonDocument>
 #include <QJsonObject>
 #include <QMetaEnum>
-#include <QTcpSocket>
 #include <functional>
 
+#include "connection.h"
 #include "controller.h"
 #include "leakdetector.h"
 #include "localizer.h"
@@ -28,172 +29,18 @@ namespace {
 
 Logger logger("Connection");
 
-struct RequestType {
-  QString m_name;
-  std::function<QJsonObject(const QJsonObject&)> m_callback;
-};
-
-void serializeServerCountry(ServerCountryModel* model, QJsonObject& obj) {
-  QJsonArray countries;
-
-  for (const ServerCountry& country : model->countries()) {
-    QJsonObject countryObj;
-    countryObj["name"] = country.name();
-    countryObj["code"] = country.code();
-
-    QJsonArray cities;
-    for (const QString& cityName : country.cities()) {
-      const ServerCity& city = model->findCity(country.code(), cityName);
-      if (!city.initialized()) {
-        continue;
-      }
-      QJsonObject cityObj;
-      cityObj["name"] = city.name();
-      cityObj["code"] = city.code();
-      cityObj["latitude"] = city.latitude();
-      cityObj["longitude"] = city.longitude();
-
-      QJsonArray servers;
-      for (const QString& pubkey : city.servers()) {
-        const Server& server = model->server(pubkey);
-        if (!server.initialized()) {
-          continue;
-        }
-
-        QJsonObject serverObj;
-        serverObj["hostname"] = server.hostname();
-        serverObj["ipv4_gateway"] = server.ipv4Gateway();
-        serverObj["ipv6_gateway"] = server.ipv6Gateway();
-        serverObj["weight"] = (double)server.weight();
-
-        const QString& socksName = server.socksName();
-        if (!socksName.isEmpty()) {
-          serverObj["socksName"] = socksName;
-        }
-
-        uint32_t multihopPort = server.multihopPort();
-        if (multihopPort) {
-          serverObj["multihopPort"] = (double)multihopPort;
-        }
-
-        servers.append(serverObj);
-      }
-
-      cityObj["servers"] = servers;
-      cities.append(cityObj);
-    }
-
-    countryObj["cities"] = cities;
-    countries.append(countryObj);
-  }
-
-  obj["countries"] = countries;
-}
-
-QJsonObject serializeStatus() {
-  MozillaVPN* vpn = MozillaVPN::instance();
-
-  QJsonObject locationObj;
-  locationObj["exit_country_code"] = vpn->serverData()->exitCountryCode();
-  locationObj["exit_city_name"] = vpn->serverData()->exitCityName();
-  locationObj["entry_country_code"] = vpn->serverData()->entryCountryCode();
-  locationObj["entry_city_name"] = vpn->serverData()->entryCityName();
-
-  QJsonObject obj;
-  obj["authenticated"] = App::isUserAuthenticated();
-  obj["location"] = locationObj;
-
-  {
-    int stateValue = vpn->state();
-    if (stateValue > App::StateCustom) {
-      MozillaVPN::CustomState state =
-          static_cast<MozillaVPN::CustomState>(stateValue);
-      const QMetaObject* meta = qt_getEnumMetaObject(state);
-      int index = meta->indexOfEnumerator(qt_getEnumName(state));
-      obj["app"] = meta->enumerator(index).valueToKey(state);
-    } else {
-      App::State state = static_cast<App::State>(stateValue);
-      const QMetaObject* meta = qt_getEnumMetaObject(state);
-      int index = meta->indexOfEnumerator(qt_getEnumName(state));
-      obj["app"] = meta->enumerator(index).valueToKey(state);
-    }
-  }
-
-  {
-    Controller::State state = vpn->controller()->state();
-    const QMetaObject* meta = qt_getEnumMetaObject(state);
-    int index = meta->indexOfEnumerator(qt_getEnumName(state));
-    obj["vpn"] = meta->enumerator(index).valueToKey(state);
-  }
-
-  return obj;
-}
-
-static QList<RequestType> s_types{
-    RequestType{"activate",
-                [](const QJsonObject&) {
-                  MozillaVPN::instance()->activate();
-                  return QJsonObject();
-                }},
-
-    RequestType{"servers",
-                [](const QJsonObject&) {
-                  QJsonObject servers;
-                  serializeServerCountry(
-                      MozillaVPN::instance()->serverCountryModel(), servers);
-
-                  QJsonObject obj;
-                  obj["servers"] = servers;
-                  return obj;
-                }},
-
-    RequestType{"disabled_apps",
-                [](const QJsonObject&) {
-                  QJsonArray apps;
-                  for (const QString& app :
-                       SettingsHolder::instance()->vpnDisabledApps()) {
-                    apps.append(app);
-                  }
-
-                  QJsonObject obj;
-                  obj["disabled_apps"] = apps;
-                  return obj;
-                }},
-
-    RequestType{"status",
-                [](const QJsonObject&) {
-                  QJsonObject obj;
-                  obj["status"] = serializeStatus();
-                  return obj;
-                }},
-};
-
 }  // namespace
 
 namespace WebExtension {
 
-Connection::Connection(QObject* parent, QTcpSocket* connection)
+Connection::Connection(QObject* parent, QIODevice* connection)
     : QObject(parent), m_connection(connection) {
   MZ_COUNT_CTOR(Connection);
-
-#if !defined(MZ_ANDROID) && !defined(MZ_IOS)
-  // `::ffff:127.0.0.1` is the IPv4 localhost address written with the IPv6
-  // notation.
-  Q_ASSERT(connection->localAddress() == QHostAddress("::ffff:127.0.0.1") ||
-           connection->localAddress() == QHostAddress::LocalHost ||
-           connection->localAddress() == QHostAddress::LocalHostIPv6);
-#endif
 
   logger.debug() << "New connection received";
 
   Q_ASSERT(m_connection);
   connect(m_connection, &QTcpSocket::readyRead, this, &Connection::readData);
-
-  MozillaVPN* vpn = MozillaVPN::instance();
-
-  connect(vpn, &MozillaVPN::stateChanged, this, &Connection::writeState);
-  connect(vpn->controller(), &Controller::stateChanged, this,
-          &Connection::writeState);
 }
 
 Connection::~Connection() {
@@ -250,6 +97,10 @@ void Connection::readData() {
   }
 }
 
+void Connection::writeMessage(QJsonObject& data) {
+  writeData(QJsonDocument(data).toJson(QJsonDocument::Compact));
+}
+
 void Connection::writeData(const QByteArray& data) {
   uint32_t length = (uint32_t)data.length();
   char* rawLength = reinterpret_cast<char*>(&length);
@@ -260,12 +111,6 @@ void Connection::writeData(const QByteArray& data) {
   }
 }
 
-void Connection::writeState() {
-  QJsonObject obj = serializeStatus();
-  obj["t"] = "status";
-
-  writeData(QJsonDocument(obj).toJson(QJsonDocument::Compact));
-}
 
 void Connection::writeInvalidRequest() {
   QJsonObject obj;
@@ -281,18 +126,7 @@ void Connection::processMessage(const QByteArray& message) {
   }
 
   QJsonObject obj = json.object();
-  QString typeName = obj["t"].toString();
-
-  for (const RequestType& type : s_types) {
-    if (typeName == type.m_name) {
-      QJsonObject responseObj = type.m_callback(obj);
-      responseObj["t"] = typeName;
-      writeData(QJsonDocument(responseObj).toJson(QJsonDocument::Compact));
-      return;
-    }
-  }
-
-  writeInvalidRequest();
+  emit onMessage(obj);
 }
 
 }  // namespace WebExtension
