+++ src/webextension/server.cpp
@@ -5,6 +5,7 @@
 #include "./server.h"
 
 #include <QHostAddress>
+#include <QJsonObject>
 #include <QTcpSocket>
 
 #include "./connection.h"
@@ -19,8 +20,9 @@ constexpr int SERVER_PORT = 8754;
 
 namespace WebExtension {
 
-Server::Server() {
+Server::Server(BaseAdapter* adapter) {
   MZ_COUNT_CTOR(Server);
+  m_adapter = adapter;
 
   logger.debug() << "Creating the server";
 
@@ -36,10 +38,22 @@ Server::~Server() { MZ_COUNT_DTOR(Server); }
 
 void Server::newConnectionReceived() {
   QTcpSocket* child = nextPendingConnection();
-  Q_ASSERT(child);
+
+  // `::ffff:127.0.0.1` is the IPv4 localhost address written with the IPv6
+  // notation.
+  if (child->localAddress() != QHostAddress("::ffff:127.0.0.1") &&
+      child->localAddress() != QHostAddress::LocalHost &&
+      child->localAddress() != QHostAddress::LocalHostIPv6) {
+    child->close();
+    return;
+  }
 
   Connection* connection = new Connection(this, child);
   connect(child, &QTcpSocket::disconnected, connection, &QObject::deleteLater);
+  connect(connection, &Connection::onMessage, m_adapter,
+          &BaseAdapter::onMessage);
+  connect(m_adapter, &BaseAdapter::onOutgoingMessage, connection,
+          &Connection::writeMessage);
 }
 
 }  // namespace WebExtension
